requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
google-api-python-client
httpx
openai
pydantic
redis
requests